Fix run script exception handling

+++ ScriperSol/Scriper.UnitTests/Assets/math.js
@@ -3,8 +3,8 @@ function multiply(a,b){
 }
 log("Count")
 var a = 5+4;
-log(a, "5+4={0}");
+logf(a, "5+4={0}");
 var b = 7-2;
-log(a, "7-2={0}");
+logf(a, "7-2={0}");
 var result = multiply(a,b);
-log(result, "multiply(a,b)={0}");
+logf(result, "multiply(a,b)={0}");
